+++ telegrambots-abilities/src/main/java/org/telegram/abilitybots/api/objects/ReplyFlow.java
@@ -1,6 +1,7 @@
 package org.telegram.abilitybots.api.objects;
 
 import org.jetbrains.annotations.NotNull;
+import org.telegram.abilitybots.api.bot.BaseAbilityBot;
 import org.telegram.abilitybots.api.db.DBContext;
 import org.telegram.abilitybots.api.util.AbilityUtils;
 import org.telegram.telegrambots.meta.api.objects.Update;
@@ -10,6 +11,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Stream;
@@ -20,7 +22,7 @@ public class ReplyFlow extends Reply {
 
   private final Set<Reply> nextReplies;
 
-  private ReplyFlow(List<Predicate<Update>> conditions, Consumer<Update> action, Set<Reply> nextReplies, String name) {
+  private ReplyFlow(List<Predicate<Update>> conditions, BiConsumer<BaseAbilityBot, Update> action, Set<Reply> nextReplies, String name) {
     super(conditions, action, name);
     this.nextReplies = nextReplies;
   }
@@ -48,7 +50,7 @@ public class ReplyFlow extends Reply {
     private final DBContext db;
     private final int id;
     private List<Predicate<Update>> conds;
-    private Consumer<Update> action;
+    private BiConsumer<BaseAbilityBot, Update> action;
     private Set<Reply> nextReplies;
     private String name;
 
@@ -63,7 +65,16 @@ public class ReplyFlow extends Reply {
       this(db, replyCounter.getAndIncrement());
     }
 
+    /**
+     * @deprecated Please use {@link #action(BiConsumer)}
+     */
+    @Deprecated
     public ReplyFlowBuilder action(Consumer<Update> action) {
+      this.action = (bot, update) -> action.accept(update);
+      return this;
+    }
+
+    public ReplyFlowBuilder action(BiConsumer<BaseAbilityBot, Update> action) {
       this.action = action;
       return this;
     }
@@ -80,7 +91,7 @@ public class ReplyFlow extends Reply {
 
     public ReplyFlowBuilder next(Reply nextReply) {
       List<Predicate<Update>> statefulConditions = toStateful(nextReply.conditions());
-      Consumer<Update> statefulAction = nextReply.action().andThen(upd -> {
+      BiConsumer<BaseAbilityBot, Update> statefulAction = nextReply.action().andThen((unused, upd) -> {
         Long chatId = AbilityUtils.getChatId(upd);
         db.<Long, Integer>getMap(STATES).remove(chatId);
       });
@@ -100,16 +111,16 @@ public class ReplyFlow extends Reply {
 
     public ReplyFlow build() {
       if (action == null)
-        action = upd -> {};
+        action = (bot, upd) -> {};
 
-      Consumer<Update> statefulAction;
+      BiConsumer<BaseAbilityBot, Update> statefulAction;
       if (nextReplies.size() > 0) {
-        statefulAction = action.andThen(upd -> {
+        statefulAction = action.andThen((bot, upd) -> {
           Long chatId = AbilityUtils.getChatId(upd);
           db.<Long, Integer>getMap(STATES).put(chatId, id);
         });
       } else {
-        statefulAction = action.andThen(upd -> {
+        statefulAction = action.andThen((bot, upd) -> {
           Long chatId = AbilityUtils.getChatId(upd);
           db.<Long, Integer>getMap(STATES).remove(chatId);
         });