+++ telegrambots-abilities/src/test/java/org/telegram/abilitybots/api/bot/AbilityBotTest.java
@@ -40,7 +40,7 @@ import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 import static org.mockito.internal.verification.VerificationModeFactory.times;
-import static org.telegram.abilitybots.api.bot.DefaultBot.getDefaultBuilder;
+import static org.telegram.abilitybots.api.bot.DefaultBot.*;
 import static org.telegram.abilitybots.api.bot.TestUtils.CREATOR;
 import static org.telegram.abilitybots.api.bot.TestUtils.*;
 import static org.telegram.abilitybots.api.db.MapDBContext.offlineInstance;
@@ -210,7 +210,7 @@ public class AbilityBotTest {
     // Support for null parameter matching since due to mocking API changes
     when(sender.downloadFile(ArgumentMatchers.<File>isNull())).thenReturn(backupFile);
 
-    defaultAbs.recoverDB().replies().get(0).actOn(update);
+    defaultAbs.recoverDB().replies().get(0).actOn(bot, update);
 
     verify(silent, times(1)).send(RECOVER_SUCCESS, GROUP_ID);
     assertEquals(db.getSet(TEST), newHashSet(TEST), "Bot recovered but the DB is still not in sync");
@@ -234,8 +234,8 @@ public class AbilityBotTest {
 
     defaultAbs.demoteAdmin().action().accept(context);
 
-    Set<Integer> actual = bot.admins();
-    Set<Integer> expected = emptySet();
+    Set<Long> actual = bot.admins();
+    Set<Long> expected = emptySet();
     assertEquals(expected, actual, "Could not sudont super-admin");
   }
 
@@ -247,8 +247,8 @@ public class AbilityBotTest {
 
     defaultAbs.promoteAdmin().action().accept(context);
 
-    Set<Integer> actual = bot.admins();
-    Set<Integer> expected = newHashSet(USER.getId());
+    Set<Long> actual = bot.admins();
+    Set<Long> expected = newHashSet(USER.getId());
     assertEquals(expected, actual, "Could not sudo user");
   }
 
@@ -259,8 +259,8 @@ public class AbilityBotTest {
 
     defaultAbs.banUser().action().accept(context);
 
-    Set<Integer> actual = bot.blacklist();
-    Set<Integer> expected = newHashSet(USER.getId());
+    Set<Long> actual = bot.blacklist();
+    Set<Long> expected = newHashSet(USER.getId());
     assertEquals(expected, actual, "The ban was not emplaced");
   }
 
@@ -273,8 +273,8 @@ public class AbilityBotTest {
 
     defaultAbs.unbanUser().action().accept(context);
 
-    Set<Integer> actual = bot.blacklist();
-    Set<Integer> expected = newHashSet();
+    Set<Long> actual = bot.blacklist();
+    Set<Long> expected = newHashSet();
     assertEquals(expected, actual, "The ban was not lifted");
   }
 
@@ -290,8 +290,8 @@ public class AbilityBotTest {
 
     defaultAbs.banUser().action().accept(context);
 
-    Set<Integer> actual = bot.blacklist();
-    Set<Integer> expected = newHashSet(USER.getId());
+    Set<Long> actual = bot.blacklist();
+    Set<Long> expected = newHashSet(USER.getId());
     assertEquals(expected, actual, "Impostor was not added to the blacklist");
   }
 
@@ -308,8 +308,8 @@ public class AbilityBotTest {
 
     defaultAbs.claimCreator().action().accept(context);
 
-    Set<Integer> actual = bot.admins();
-    Set<Integer> expected = newHashSet(CREATOR.getId());
+    Set<Long> actual = bot.admins();
+    Set<Long> expected = newHashSet(CREATOR.getId());
     assertEquals(expected, actual, "Creator was not properly added to the super admins set");
   }
 
@@ -335,8 +335,8 @@ public class AbilityBotTest {
 
     bot.addUser(update);
 
-    Map<String, Integer> expectedUserIds = ImmutableMap.of(USER.getUserName(), USER.getId());
-    Map<Integer, User> expectedUsers = ImmutableMap.of(USER.getId(), USER);
+    Map<String, Long> expectedUserIds = ImmutableMap.of(USER.getUserName(), USER.getId());
+    Map<Long, User> expectedUsers = ImmutableMap.of(USER.getId(), USER);
     assertEquals(expectedUserIds, bot.userIds(), "User was not added");
     assertEquals(expectedUsers, bot.users(), "User was not added");
   }
@@ -350,15 +350,15 @@ public class AbilityBotTest {
     String newUsername = USER.getUserName() + "-test";
     String newFirstName = USER.getFirstName() + "-test";
     String newLastName = USER.getLastName() + "-test";
-    int sameId = USER.getId();
+    long sameId = USER.getId();
     User changedUser = new User(sameId, newFirstName, false, newLastName, newUsername, "en", false, false, false);
 
     mockAlternateUser(update, message, changedUser);
 
     bot.addUser(update);
 
-    Map<String, Integer> expectedUserIds = ImmutableMap.of(changedUser.getUserName(), changedUser.getId());
-    Map<Integer, User> expectedUsers = ImmutableMap.of(changedUser.getId(), changedUser);
+    Map<String, Long> expectedUserIds = ImmutableMap.of(changedUser.getUserName(), changedUser.getId());
+    Map<Long, User> expectedUsers = ImmutableMap.of(changedUser.getId(), changedUser);
     assertEquals(bot.userIds(), expectedUserIds, "User was not properly edited");
     assertEquals(expectedUsers, expectedUsers, "User was not properly edited");
   }
@@ -659,6 +659,30 @@ public class AbilityBotTest {
 
     String expected = "test channel post";
     verify(silent, times(1)).send(expected, 1);
+  }
+
+  @Test
+  void canProcessRepliesRegisteredInCollection() {
+    Update firstUpdate = mock(Update.class);
+    Message firstMessage = mock(Message.class);
+    when(firstMessage.getText()).thenReturn(FIRST_REPLY_KEY_MESSAGE);
+    when(firstMessage.getChatId()).thenReturn(1L);
+
+    Update secondUpdate = mock(Update.class);
+    Message secondMessage = mock(Message.class);
+    when(secondMessage.getText()).thenReturn(SECOND_REPLY_KEY_MESSAGE);
+    when(secondMessage.getChatId()).thenReturn(1L);
+
+    mockUser(firstUpdate, firstMessage, USER);
+    mockUser(secondUpdate, secondMessage, USER);
+
+
+    bot.onUpdateReceived(firstUpdate);
+    bot.onUpdateReceived(secondUpdate);
+
+    verify(silent, times(2)).send(anyString(), anyLong());
+    verify(silent, times(1)).send("first reply answer", 1);
+    verify(silent, times(1)).send("second reply answer", 1);
   }
 
   private void handlesAllUpdates(Consumer<Update> utilMethod) {