+++ telegrambots-abilities/src/test/java/org/telegram/abilitybots/api/bot/DefaultBot.java
@@ -5,6 +5,7 @@ import org.telegram.abilitybots.api.objects.Ability;
 import org.telegram.abilitybots.api.objects.Ability.AbilityBuilder;
 import org.telegram.abilitybots.api.objects.Flag;
 import org.telegram.abilitybots.api.objects.Reply;
+import org.telegram.abilitybots.api.objects.ReplyCollection;
 import org.telegram.abilitybots.api.toggle.AbilityToggle;
 
 import static org.telegram.abilitybots.api.objects.Ability.builder;
@@ -15,6 +16,8 @@ import static org.telegram.abilitybots.api.objects.Privacy.ADMIN;
 import static org.telegram.abilitybots.api.objects.Privacy.PUBLIC;
 
 public class DefaultBot extends AbilityBot {
+  public static final String FIRST_REPLY_KEY_MESSAGE = "first reply key string";
+  public static final String SECOND_REPLY_KEY_MESSAGE = "second reply key string";
 
   public DefaultBot(String token, String username, DBContext db) {
     super(token, username, db);
@@ -35,16 +38,16 @@ public class DefaultBot extends AbilityBot {
   }
 
   @Override
-  public int creatorId() {
-    return 1337;
+  public long creatorId() {
+    return 1337L;
   }
 
   public Ability defaultAbility() {
     return getDefaultBuilder()
         .name(DEFAULT)
         .info("dis iz default command")
-        .reply(Reply.of(upd -> silent.send("reply", upd.getMessage().getChatId()), MESSAGE, update -> update.getMessage().getText().equals("must reply")).enableStats("mustreply"))
-        .reply(upd -> silent.send("reply", upd.getCallbackQuery().getMessage().getChatId()), CALLBACK_QUERY)
+        .reply(Reply.of((bot, upd) -> silent.send("reply", upd.getMessage().getChatId()), MESSAGE, update -> update.getMessage().getText().equals("must reply")).enableStats("mustreply"))
+        .reply((bot, upd) -> silent.send("reply", upd.getCallbackQuery().getMessage().getChatId()), CALLBACK_QUERY)
         .build();
   }
 
@@ -75,7 +78,20 @@ public class DefaultBot extends AbilityBot {
 
   public Reply channelPostReply() {
     return Reply.of(
-        upd -> silent.send("test channel post", upd.getChannelPost().getChatId()), Flag.CHANNEL_POST
+            (bot, upd) -> silent.send("test channel post", upd.getChannelPost().getChatId()), Flag.CHANNEL_POST
+    );
+  }
+
+  public ReplyCollection createReplyCollection() {
+    return ReplyCollection.of(
+        Reply.of(
+            upd -> silent.send("first reply answer", upd.getMessage().getChatId()),
+            update -> update.getMessage().getText().equalsIgnoreCase(FIRST_REPLY_KEY_MESSAGE)
+        ),
+        Reply.of(
+            upd -> silent.send("second reply answer", upd.getMessage().getChatId()),
+            update -> update.getMessage().getText().equalsIgnoreCase(SECOND_REPLY_KEY_MESSAGE)
+        )
     );
   }